Move skills endpoint into integrations due deprecation with backward compatibility
